slaweb_api/requirements.txt
django==1.6
psycopg2>=2.5.1
psycogreen==1.0
djangorestframework>=2.3.9,<=2.3.14
djangorestframework-jwt>=0.1.2
django-extensions>=1.2.5
django-cors-headers>=0.12
gevent-socketio==0.3.6
simplejson>=3.0.8
jsonfield>=0.9.20
suds>=0.4
coverage>=3.7.1
gunicorn==18.0
django-imagekit>=3.2
pilkit>=1.1.6
reportlab==3.1.8
celery>=3.1.12
Pillow>=2.4.0
redis>=2.10.1
pymongo>=2.7.2,<=2.8
netaddr>=0.7.13
django-celery==3.1.16
python-dateutil==1.5
Fabric>=1.10.0